pyproject.toml
[tool.poetry]
name = "pytorch-ssd"
version = "0.1.0"
description = ""
authors = ["AJ <akarsh.1995.02@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.6.1,<4"
torch = "^1.7.0"
torchvision = "^0.8.1"
pandas = "^1.1.4"
opencv-python = "^4.4.0"
install = "^1.3.4"
pascal-voc-writer = "^0.1.4"
pytesseract = "^0.3.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
